requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0
networkx>=3.1
python-louvain>=0.16

# Database and Storage
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=4.5.0
aioredis>=2.0.0
pymongo>=4.3.0

# Google Gemini AI
google-generativeai>=0.3.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
aiohttp>=3.8.5
backoff>=2.2.1

# Data Validation and Quality
pydantic-settings>=2.0.0
marshmallow>=3.19.0
cerberus>=1.3.4

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
tqdm>=4.65.0
requests>=2.31.0

# Visualization and Monitoring
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.4.0

# Recommendation Engine Dependencies
implicit>=0.7.0
scikit-surprise>=1.1.3
lightfm>=1.17
faiss-cpu>=1.7.4
joblib>=1.3.0
numba>=0.57.0
python-rapidjson>=1.10
cachetools>=5.3.0

# NLP and Language Processing
spacy>=3.6.0
langdetect>=1.0.9

# Model Training and Evaluation
optuna>=3.2.0
mlflow>=2.4.0
tensorboard>=2.13.0

# Streaming and Adaptive Learning
kafka-python>=2.0.2
redis-py-cluster>=2.1.0
river>=0.18.0
confluent-kafka>=2.3.0
asyncio-mqtt>=0.13.0
websockets>=11.0.2

# Online Learning and Drift Detection
scikit-multiflow>=0.5.3
concept-drift-detector>=1.0.0

# Additional Phase 3 Dependencies
aioredis>=2.0.1
dash>=2.14.0
statsmodels>=0.14.0
pykalman>=0.9.5
memory-profiler>=0.61.0
line-profiler>=4.1.0
colorlog>=6.8.0
starlette>=0.27.0
aiosqlite>=0.19.0
alembic>=1.12.0
asyncpg>=0.29.0
aiofiles>=23.2.0
click>=8.1.7
python-dateutil>=2.8.2
pytz>=2023.3
jsonschema>=4.20.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Advanced Analytics
dataclasses-json>=0.6.1
typing-extensions>=4.7.0

# Production Deployment Dependencies
uvloop>=0.19.0
httptools>=0.6.0
python-multipart>=0.0.6
prometheus-client>=0.19.0
sentry-sdk>=1.40.0

# Demo Interface Dependencies
streamlit>=1.28.0
plotly>=5.17.0
altair>=5.2.0
Pillow>=10.1.0

# Security and Authentication
PyJWT>=2.8.0
cryptography>=41.0.0
bcrypt>=4.1.0

# Monitoring and Alerting
psutil>=5.9.0
scikit-learn>=1.3.0

# Load Testing
locust>=2.17.0

# Additional ML Dependencies
concept-drift-detector>=1.0.0
